feat: Add support for font icons in GUI
The code changes in this commit add support for using font icons in the GUI. This feature is controlled by the `FONT_ICONS` flag in the `GlobalSettings` class. When enabled, the GUI will display organism types as text icons instead of image icons. This change improves the flexibility and customization options for the GUI.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (feat for feature).